Adds a very simple js client
- The client itself is created with two parameters, one for the register
url of the RP and one for the login url of the RP.

- When calling register it does a GET towards the RP for getting all the
data necessary for creating a key pair and hands over responsibility to
the authenticator. Whatever comes back is then sent back to the RP in
the expected format by making a POST to the same endpoint.

- The login works in similar wat by making a GET request for getting
the necessary data to perform login. That is handed over to the
authenticator and what comes back is sent to the same endpoint with
POST.

All in all, this client assumes the RP listens for GET and POST on the
same url.